Initial work on iOS app
This is a working build of the Soundbounce iOS app. All it does right
now is load a UIWebView pointing to the web login, but I've linked up a
JavaScript interface which will be used to implement the Spotify library
connection.
Also in this commit is a tiny "fix" to stop the roomlist showing a 404
error in the console for rooms which don't have an image.